Remove the distinction between a driver and a pool
The pool now maintains pointers to driver implementation callbacks.

No longer pass a config arg to each driver callback, just the OpenProc. It
is the Open callback's responsibility to stash the config data in the handle
client-data arg.

Also, add a client data arg to to the Statement structure and make a lot
of it's other members part of a public interface, no longer passing them
explicitly to the PrepareProc callback.