Add Dbi_ColumnLength and Dbi_ColumnLengthProc for sizing buffers
Some dbs require you to pass them a buffer of the correct size into
which they will copy a column value, whereas other will pass you
a pointer to the data which your are responsible for copying out.

This adds support for the first scheme and allows some drivers to
skip an extra copy.